refactor: allow some null values, and add some extra tests
Mostly modified some of the setters for nullable values so that null
can also be passed. This just makes it easier on the consuming
side where you have a value like `?string` that you wanna set.
Better to allow null there, rather than expect consumers to add
extra conditionals to deal with it.

Also updated the document class so it includes getters and setters
(as opposed to just additive methods) to make it more flexible on
the consumer side. If you can build it all in one pass and wanna
modify it with some kind of post processor it was basically
impossible before; the getters/setters make this easier.

Added a bunch of extra tests.

Removed tests directory from phpstan analysis. Seemed like a bad
idea, but thought i'd give it a go. It's a lot of pain keeping
all the stubs satisfying that level of stan.